selenium_tests/requirements.txt
httpx==0.23.0
webdriver-manager==3.8.2
selenium==4.3.0
pytest==7.0.1
pytest-bdd==6.0.1
pytest-asyncio==0.18.2
PyHamcrest==2.0.3
